Step 1 Refactoring
The chargeForRental(Rental each) method is extracted from statement() method. The Rental object is passed as a parameter and the chargeForRental method is called in each iteration the rest switch statement code is kept same, other than introducing a new temporary variable named 'amount', in which the individual rent charge is calculated and stored and then returned to the statement() method.