Step 3 Refactoring
In this step the method chargeForRental() is moved from Customer.java to Rental.java, is renamed to charge() and is made public instead of private. All the references has been changed as all the functions are of Rental.java class only. In Customer.java class the function is called as rental.charge(), and there are also some variable name changes in the Customer.java class.